Update README. Make modification optional.

+++ bootimg.py
@@ -22,13 +22,26 @@ def repack_boot(base_dir, tmp_dir, out):
     p = subprocess.Popen([mkboot, tmp_dir, join(out, 'newboot.img')])
     p.communicate()
 
-def unpack_repack(bootimg, base_dir, tmp_dir, out, sepolicy):
+def unpack_repack(bootimg, base_dir, tmp_dir, out, sepolicy, file_contexts, seapp_contexts, property_contexts):
 
     ramdisk_dir = unpack_boot(bootimg, base_dir, tmp_dir)
     
-    #copy sepolicy
-    copy(sepolicy, ramdisk_dir)
+    if sepolicy and isfile(sepolicy):
+        print 'Copying sepolicy'
+        copy(sepolicy, ramdisk_dir)
 
+    if file_contexts and isfile(file_contexts):
+        print 'Copying file_contexts'
+        copy(file_contexts, ramdisk_dir)
+
+    if seapp_contexts and isfile(seapp_contexts):
+        print 'Copying seapp_contexts file'
+        copy(seapp_contexts, ramdisk_dir)
+
+    if property_contexts and isfile(property_contexts):
+        print 'Copying property_contexts file'
+        copy(property_contexts, ramdisk_dir)
+        
     repack_boot(base_dir, tmp_dir, out)
 
 
@@ -45,7 +58,12 @@ def update_boot(args):
     
     makedirs(args.out_dir)
 
-    unpack_repack(args.bootimg, project_dir, tmp_dir, args.out_dir, args.sepolicy)
+    if not isfile(args.bootimg):
+        print 'Please specify a valid boot.img file'
+        raise SystemExit
+
+    unpack_repack(args.bootimg, project_dir, tmp_dir, args.out_dir, args.sepolicy,
+                  args.file_contexts, args.seapp_contexts, args.property_contexts)
     rmtree(tmp_dir)
 
 def main():
@@ -53,7 +71,7 @@ def main():
 
     parser = argparse.ArgumentParser(description='Unpack and repack boot.img')
     parser.add_argument('-boot', required=True, dest='bootimg', help='boot.img file')
-    parser.add_argument('-sepolicy', required=True, dest='sepolicy', 
+    parser.add_argument('-sepolicy', required=False, dest='sepolicy', 
                         help='new sepolicy file to add in the boot.img')
     parser.add_argument('-file_contexts', required=False, dest='file_contexts', 
                         help='new file_contexts file to add in the boot.img')
